Add -nr option to exclude local requires
Class systems that return the class table from require produce lots of
noise but the class definition is removed by -nv. Add a new option
specifically for requires: -nr.

Example:

	local Animal = class(function(a,name)
	   a.name = name
	end)

	return Animal

	local Animal = require("Animal")

We want the first Animal (the singular definition of the class), but not
the second (the repeated import of the class).